refactor(cli): merge the 5-file instruction-builder cluster into one file
TERMINAL_RUN_GUIDANCE, formatCliRunFileInstruction,
formatToolUseAgentRunInstruction, formatUnavailableApprovalInstruction,
and formatMultiAgentRunInstruction were split across 5 files (150 lines
total) with no independent reuse axis — every symbol is consumed only
by agentsRun.ts, multiAgent.ts, and one test, plus each other. Collapse
into commands/_helpers/runInstructions.ts, following the same per-file
consolidation dispatch.ts already documents for this directory.

TERMINAL_RUN_GUIDANCE is no longer exported: it now has no consumer
outside this file.
