--- packages/cli/src/commands/_helpers/terminalRunInstruction.ts
@@ -1,2 +1,0 @@
-export const TERMINAL_RUN_GUIDANCE =
-  'This CLI run exits after your final response. Do not end by asking the user whether to perform more work; either complete the requested work now or state the exact artifacts and next command/action for a future run.';
